Fix algo size bug
The function _create_algo_bin uses the size of sect_ro rather than the
current sector size when generating the algo. This causing the final
size of the algo to be wrong. This patch updates that code to instead
use the current section size. It also asserts that the size of data is
equal to the size assigned. This prevents the algo_data array from
unexpectedly changing size.